chore: move internal to pkg
